qwertyqwerty

+++ app/src/main/java/uz/texnopos/installment/data/model/Client.kt
@@ -5,12 +5,12 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Client(
-    val all_sum: Int,
+    val all_sum: Double,
     val client_id: Int,
     val client_name: String,
     val count: Int,
     val latter: String,
-    val paid: Int,
+    val paid: Double,
     val pasport_number: String,
     val pasport_photo: String,
     val pasport_serial: String,
